(PC-4996) refacto: move Spacer to avoid require loop
